Better way to do css processing
