--- app.css
@@ -1,6 +1,0 @@
-@import "tailwindcss";
-@plugin "@tailwindcss/typography";
-@source "./src/**/*.{html,md,liquid,njk,11ty.js}";
-@plugin "daisyui" {
-    logs: false;
-}
