+++ package.json
@@ -5,10 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "build": "npx @11ty/eleventy && tailwindcss -i app.css -o ./_site/assets/stylesheets/output.css --minify",
-    "dev:serve": "npx @11ty/eleventy --serve",
-    "dev:tailwind": "tailwindcss -i app.css -o ./_site/assets/stylesheets/output.css --watch",
-    "start": "run-p dev:serve dev:tailwind"
+    "build": "eleventy",
+    "start": "eleventy --serve"
   },
   "repository": {
     "type": "git",
@@ -19,14 +17,15 @@
     "url": "https://github.com/UmbreoClaw/MyWebsite/issues"
   },
   "homepage": "https://umbreoclaw.net",
-  "devDependencies": {
+  "dependencies": {
     "@11ty/eleventy": "^3.0.0",
+    "@tailwindcss/postcss": "^4.1.3",
     "@tailwindcss/typography": "^0.5.16",
+    "cssnano": "^7.0.6",
     "daisyui": "^5.0.12",
     "html-minifier-terser": "^7.2.0",
-    "npm-run-all2": "7.0.2",
-    "tailwindcss": "^4.1.2",
-    "@tailwindcss/cli": "^4.1.2"
+    "postcss": "^8.5.3",
+    "tailwindcss": "^4.1.3"
   },
   "engines": {
     "node": ">=18"
